pyproject.toml
[tool.poetry]
name = 'lightnovel-crawler'
version = "3.0.0-alpha"
description = "An app that downloads novels and generate e-books from online sources."
authors = ["Sudipto Chandra <dipu.sudipta@gmail.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.6"
cloudscraper = "^1.2.40"
beautifulsoup4 = "^4.9.1"
html5lib = "^1.0.1"
lxml = "^4.5.1"
python-slugify = "^4.0.0"
ebooklib = "^0.17.1"
pyyaml = "^5.3.1"
js2py = "^0.70"
atomicwrites = "^1.4.0"
cairosvg = { version = "^2.4.2", optional = true }
cleo = "^0.8.1"

[tool.poetry.dev-dependencies]
"discord.py" = "^1.3.3"
python-telegram-bot = "^12.7"
flake8 = "^3.8.3"
mypy = "^0.780"
tox = "^3.15.2"
pytest = "^5.4.3"
pytest-env = "^0.6.2"
pytest-watch = "^4.2.0"
pytest-timeout = "^1.4.1"
pytest-forked = "^1.1.3"
pytest-xdist = "^1.32.0"
pytest-cov = "^2.10.0"
coverage = "^5.1"
coveralls = "^2.0.0"
responses = "^0.10.15"
pyinstaller = "^3.6"
autopep8 = "^1.5.3"
wheel = "^0.34.2"

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
